Add experimental features keywords and ^
1. Allow FILTER keywords(...) in the query and rewrite it using
ql:contains-word and ql:contains entity. This enables fast keyword
search in entity names, assuming that the index was build with a text
part that contains one text record per entity name which contains the
entity as entity and the name of the entity as text

1. Add basic support for the suggestion of reverse predicates (prefixed
with ^). This is realized via an added column ?qleverui_reversed in the
results returned by the predicate suggestion query. If the value in that
column is 0, it's an ordinary predicate suggestion (like so far). If the
value is 1, it's a reverse predicate suggestion, which is displayed as
the predicate name prefixed with a ^