Add settings_secret.py to simplify installation
The previous setup was more secure, but the QLever UI is not really a
high-security applications, at least for now. Anyway, the key in
settings_secret.py can be changed anytime if so desired. It's just
easier to forget this now. In my understanding, the only reason for the
previous setup was that one could not forget it because one needed to
explicitly copy the file before building the docker image.