Add button for MapView++ and remove -proxy for TSV/CSV export
1. For certain backends, there is now a button for the standad
"MapView", as well as for our new "MapView++" (which can even display
very large number of objects in no time).

2. When exporting CSV or TSV, remove "-proxy" from the API URL. This is
a hack, relevant only for our internal use (Wikidat runs via a proxy,
but when exporting CSV or TSV, we don't want this detour and the
possible timeouts that go along with it).