Prepare "name service" rewriting
The "name service" rewriting of queries (adding "name" triples
automatically, so far done by qlever-proxy.py) issues queries to the
backed and must hence be an async function. As a consequence, all
functions calling this (and the functions calling them, etc.) must be
async, to. This commit does this.

Also, much of the code for the actual "name service" is already there,
too.