Call name service only when needed + fix "show all" with LIMIT
The name service calls queries in the background and can take some time.
It should only be called when needed (which currently is only, when the
respective checkbox in the UI is checked).

On the side, fix an issue brought up in
https://github.com/ad-freiburg/qlever/discussions/713 : don't show the
"show all" button when the query has a LIMIT X and all X rows are shown
(even if without the LIMIT there would be more results).