On mouseover, issue name query only if actually defined
On mouseover over a subject, predicate, or object, the QLever UI issues
a name query (which can be specified in the backend config) and if
successful, shows the name as a tooltip.

So far, there was a check if *any* of these three queries were specified
and if yes, the name query was issued for subject, predicate, and
object. Now the query is issued only if that particular query has indeed
been specified in the config. This fixes a bug, where many empty queries
were issued when only some of the name queries are defined in the
config.