Update AC templates to reflect PR #795 from the qlever repo
With https://github.com/ad-freiburg/qlever/pull/795, the current AC
queries (which did the right thing, but techically used the wrong
syntax, in particular no DISTINCT in conjunction with ql:has-predicate)
no longer work. The new AC templates now work with the latest version of
Qlever. See PR #795 for details.

TODO: In a next step, rewrite the AC queries such that they don't use
ql:has-predicate explicitly (but leave it to the QLever backend to
figure this out itself, which after PR #795 it is capable of doing).