Highlight tokens causing parse error (#51)
Since https://github.com/ad-freiburg/qlever parses the SPARQL query
using ANTLR (this was a series of multiple PRs), the error result
returned, when a parse error occurs, contains information about the
exact span of the parse error in the SPARQL query. This change makes use
of that information to highlight the respective tokens.