Refactor function getFormattedResultEntry
This function was called getShortStr. It takes an element from a result
table (element = entry at a particular row and column) and converts it
to the string that is then displayed in the QLever UI. This also does
some additional magic, like adding icons to the column header if the
element has a particular datatype. The function is now much cleaner,
less buggy, and has a more appropriate name.