Improve function rewriteQuery for ogc:contains
Refactored the code, so that it is now cleaner and can also handle
ogc:contain rewrites for queries with UNION, MINUS, and OPTIONAL.

Also add code for rewriting SERVICE clauses. This was for testing
purposes only (since QLever's #793 is on the way) and is deactivated.
But I still want to leave it here, so that it's preserved.

A minor unrelated change: Show "optimized-out" status in blue. This
would have belonged into 1fd0fc.