Show information about "lazy scans" in "Analysis" view
The runtime information shown in the "Analysis" tree now contains
additional information, notably such related to "lazy index scans", see
PR https://github.com/ad-freiburg/qlever/pull/1003 .

The status of the "normal" operations has been renamed to "fully
materialized" and is not explicitly shown in the tree. The status of
the lazy index scans is called "lazily materialized".

When hovering over a node in the tree, more detailed infomation is shown
at the bottom if available.