Consistently use `fetch` + improve error handling (#69)
The JavaScript code of the QLever UI launches queries at many places (to get a query result, to get suggestions, to clear the cache, to get statistics, to get a share link, etc.). In the code so far, each request did this slightly differently. This is now unified as much as possible. In particular, all requests now use `fetch` and none of the requests use `jQuery.ajax` anymore.

Along the way, improve the error messages when something goes wrong with a `fetch` and remove code that is no longer needed (like the code that realized `SERVICE` in the JavaScript).