Add no-slug mode (#66)
So far, the URL of a backend in the QLever UI always has the form `https://qlever.cs.uni-freiburg.de/<slug>`, where `slug` is a short URL-appropriate name for the backend. This PR now allows to specify one backend, which is used when the same URL is specified, but without the `<slug>`. This backend must also be the default backend then. The  query URLs can then also have the form `https://<domain>/334GvZf`, that is, without slug. All URLs also work with the slug.

NOTE: This change requires a migration of the `qleverui.sqlite3` database because there is now a new field `isNoSlugMode` for each backend configuration. We use the occasion to update the database to the latest configuration (and example queries) for our backends.