Suggest geo:hasGeometry/geo:asWKT automagically (#74)
For a while now, the TTL files on https://osm2rdf.cs.uni-freiburg.de provide the geometry WKTs via `geo:hasGeometry/geo:asWKT` (because that is how it should be according to the GeoSPARQL standard), and not directly via `geo:hasGeometry` as it used to be.

To ease query construction, the QLever UI now automagically suggests `geo:hasGeometry/geo:asWKT` whenever `geo:hasGeometry` is among the suggested predicates. We had a similar mechanism already for predicates like `wdt:P31` (in which case also `wdt:P31/wdt:P279*` is suggested).

When `geo:hasGeometry/geo:asWKT` is selected by the user, the variable name suggestion is `?geometry` and not `?as_wkt`.

On the side, also show the Map view button for OHM (when the result contains geometries in the last column, as usual).