Add keyword search to filter example queries (#83)
When clicking on the button "Examples", there is now a search field on top of the list of example queries. This search field automatically has the focus, and when typing something, then after each keystroke, the list of example queries is reduced to those, where the description matches the search query. A description matches when each part of the search query matches a part of the description. The parts of the search query can even be regular expressions.