Clean up the .dockerignore file
NOTE: The `db` directory (which contains the database `qleverui.sqlite3`
for the web app) should *not* be included in the Docker container, but
should be mounted as a volume. It contains valuable information like the
query hashes (shortcuts for queries that users type) and backend
configurations. These change when an instance is actively used for an
extended period of time (like the main instace on
https://qlever.cs.uni-freiburg.de). If the database were only stored
inside of the containe, the risk that the changes get lost (for example,
by killing the container and restarting it) are high.