Fix links for `_sparql` and `_mapview` variable hack
When a variable name ends with `_sparql` or `_mapview` and the value
is a link, then a special link is shown which leads to an instance of
the QLever UI with a SPARQL query or to an instance of Petrimaps,
respectively.

This was broken because the link was enclosed in quotes. These are now
removed when they are there.

In collaboration with @qup42