Add sort key to example queries (#108)
Each example query now has a sort key, which determines how the example queries are ordered in the `Examples` drop-down menu and for the `examples` command. The default sort key is `~`, which is larger than most characters. Ties are broken by the `id` of the example query, which increases over time (example queries added later get a larger `id). That way, without entering new sort keys, the order is the same as before.

NOTE: The migration file `backend/migrations/0074_...` also contains migrations from earlier commits, which by mistake were not added to the repository. But better late than never.