Bound box width in "Analysis" tree
For queries with many variables, the box for a node in the "Analysis"
tree can become very wide, which makes visual parsing very hard. This
change limits the width of the fields for the operation and the column
(which both list variables) to 30 characters, and shows ellipsis if they
are longer. It also sets the `title` field, so that one can always get
the full contents in the tooltip via a mouseover.