Support updates in the QLever UI (#126)
With this change, the QLever UI supports updates. Under "Backend Information", an access token can be entered, in which case, two new buttons "Clear Delta Triples" and "Clear Cache (Complete)" appear. When an update has been entered and executed and the access token is correct and the update worked, a message "Update successful (insert triples: ..., delete triples: ...)" is displayed. Fixes #122