Rename command to `config` + first working version
The command can now be used to get the current config as YAML (call it
without argument).

Or you can set a new config by providing a YAML file as argument. With
the option `--hide-all-other-backends` one can hide all other backends
from the config.

The example queries are also included (both for getting and setting).

Setting a subset of the field is possible. As of now, the `id` and
`slug` cannot be changed.