Fix scuffed encoding of `raphael.js` library (#142)
Parts of this library were not properly UTF-8 encoded. This is now fixed by replacing it with the closest version in the official Raphaël repository: https://github.com/DmitryBaranovskiy/raphael/blob/f7744d0e44e9ef13ddab2a53eb14b16d8c875ddc/raphael-min.js and manually adjusting the code, so that the non UTF-8 related part is the same as before. It is unclear where the code used so far in the QLever UI comes from (it cannot be found in https://github.com/DmitryBaranovskiy/raphael).