New command `config` (#136)
New command to get or set the configuration for a backend from the command line, as follows:

```
python manage.py config <backend slug>
python manage.py config <backend slug> <config>.yml
```

The first command writes the current configuration (including examples queries) as YAML to stdout. The second command updates the configuration according to the specified YAML. The YAML may contain only a subset of the fields, in which case only those fields will be updated. If the format of the YAML is invalid or fields are specified that do not exist in the QLever UI configuration, no action at all will be taken. The fields `id` and `slug` cannot be changed.

The second command has an option `--hide-all-other-backends`, which has the effect that all other backends will be hidden (by setting their sort key to `0`).

When logged into the QLever UI, the YAML can also be accessed via an API call such as this one: https://qlever.cs.uni-freiburg.de/api/config/wikidata

Part of an effort to address #125