Support chained updates in the UI (#143)
For multiple updates, this shows aggregated statistics (also saying how many updates were aggregated). The query response used to be a JSON object and is now an array of JSON objects. For "Analysis", the last update in a chain is shown.

This should be merged before ad-freiburg/qlever#1883, which implemented chained updates in QLever.