Show buttons for privileged commands in main panel (#144)
When a string is entered in the "Access Token" field in the "Backend Information" panel, two buttons for privileged commands appear (one for removing all delta triples and one for clearing the cache completely, including the pinned results). So far, these buttons appeared in the same "Backend Information" panel, which is usually hidden. Now they appear in the row below the usual buttons (for executing the query, downloading the result, etc).