Remove key `id` from YAML file exported by `config` command (#148)
So far, the YAML exported or imported by the `config` command contained values for both `id` and `slug` under `backend`. Indeed, both of these are part of the `Backend` model. If the values for these differed between the YAML and the database, an error was reported. However, this made it challenging to exchange such YAML files between different users.

We now consider the `id` an implementation detail that should neither be exported nor imported. Indeed, an error is now reported if the YAML file contains an `id` key under `backend`. Only the `slug` key is used to identify the backend in the database.

NOTE: For historical reasons, the value of `slug` is not unique in the `Backend` model. It should be, but that is work for a separate PR (which will require a database migration).