Remove `@login_required` from calls to `config` API
None of the information returned by the config API is really secret.
Indeed, most of it could be obtained by interacting with the respective
instance of the QLever UI (except for the `apiToken`, which, however, is
a relic of the past and should be removed from the config altogether).

On the positive side, making this API public help users who want to
set up their own QLever UI instance. In particular, they can now simply
get a copy of the config of an existing instance and modify it to their
own liking.